supplier added and migration fix

+++ database/migrations/2022_12_15_220758_create_invoice_table.php
@@ -17,7 +17,7 @@ class CreateInvoiceTable extends Migration
             $table->id('invoice_no');
             $table->string('manual')->nullable();
             $table->timestamp('date')->nullable();
-            $table->unsignedBigInteger('customer')->nullable();
+            $table->unsignedInteger('customer')->nullable();
             $table->string('authorized')->nullable();
             $table->string('phone')->nullable();
             $table->string('trn')->nullable();
